Move 'addLinking' to Builder.hs.
