+++ cabal-install/Distribution/Solver/Modular/Linking.hs
@@ -1,9 +1,7 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 module Distribution.Solver.Modular.Linking (
-    addLinking
-  , validateLinking
+    validateLinking
   ) where
 
 import Prelude ()
@@ -30,67 +28,6 @@ import qualified Distribution.Solver.Modular.WeightedPSQ as W
 import Distribution.Solver.Types.OptionalStanza
 import Distribution.Solver.Types.PackagePath
 import Distribution.Solver.Types.ComponentDeps (Component)
-
-{-------------------------------------------------------------------------------
-  Add linking
--------------------------------------------------------------------------------}
-
-type RelatedGoals = Map (PN, I) [PackagePath]
-type Linker       = Reader RelatedGoals
-
--- | Introduce link nodes into the tree
---
--- Linking is a traversal of the solver tree that adapts package choice nodes
--- and adds the option to link wherever appropriate: Package goals are called
--- "related" if they are for the same instance of the same package (but have
--- different prefixes). A link option is available in a package choice node
--- whenever we can choose an instance that has already been chosen for a related
--- goal at a higher position in the tree. We only create link options for
--- related goals that are not themselves linked, because the choice to link to a
--- linked goal is the same as the choice to link to the target of that goal's
--- linking.
---
--- The code here proceeds by maintaining a finite map recording choices that
--- have been made at higher positions in the tree. For each pair of package name
--- and instance, it stores the prefixes at which we have made a choice for this
--- package instance. Whenever we make an unlinked choice, we extend the map.
--- Whenever we find a choice, we look into the map in order to find out what
--- link options we have to add.
-addLinking :: Tree d c -> Tree d c
-addLinking = (`runReader` M.empty) .  cata go
-  where
-    go :: TreeF d c (Linker (Tree d c)) -> Linker (Tree d c)
-
-    -- The only nodes of interest are package nodes
-    go (PChoiceF qpn gr cs) = do
-      env <- ask
-      let linkedCs = W.fromList $ concatMap (linkChoices env qpn) (W.toList cs)
-          unlinkedCs = W.mapWithKey (goP qpn) cs
-      allCs <- T.sequence $ unlinkedCs `W.union` linkedCs
-      return $ PChoice qpn gr allCs
-    go _otherwise =
-      innM _otherwise
-
-    -- Recurse underneath package choices. Here we just need to make sure
-    -- that we record the package choice so that it is available below
-    goP :: QPN -> POption -> Linker (Tree d c) -> Linker (Tree d c)
-    goP (Q pp pn) (POption i Nothing) = local (M.insertWith (++) (pn, i) [pp])
-    goP _ _ = alreadyLinked
-
-linkChoices :: forall a w . RelatedGoals
-            -> QPN
-            -> (w, POption, a)
-            -> [(w, POption, a)]
-linkChoices related (Q _pp pn) (weight, POption i Nothing, subtree) =
-    map aux (M.findWithDefault [] (pn, i) related)
-  where
-    aux :: PackagePath -> (w, POption, a)
-    aux pp = (weight, POption i (Just pp), subtree)
-linkChoices _ _ (_, POption _ (Just _), _) =
-    alreadyLinked
-
-alreadyLinked :: a
-alreadyLinked = error "addLinking called on tree that already contains linked nodes"
 
 {-------------------------------------------------------------------------------
   Validation
